Update branding and ownership information
