+++ README.md
@@ -1,37 +1,33 @@
-# Airline Reservation Management System
+# MyAirline Reservation Management System
 
-The Airline Reservation Management System is a web-based application that streamlines the process of flight booking and reservation management. It offers a convenient interface for users to search and book flights, check available seats, manage reservations, and handle cancellations. The system also provides airline administrators with tools to manage flights, seat availability, and passenger information.
-
-![Screenshot (289)](https://github.com/ImeshaDilshani/Airline-Reservation-Management-System/assets/93858302/28855982-352a-4b09-a0a5-d7aa8bcf6c5f)
-
+MyAirline Reservation Management System is a web-based application for handling flight bookings and reservations. The project was built with Java Spring Boot and offers an interface for passengers and administrators alike.
 
 ## Features
-- User Registration and Login: Passengers can create accounts and log in to access the reservation system.
-- Flight Search: Users can search for available flights based on their preferred destinations and travel dates.
-- Seat Availability: The system displays seat availability for each flight, allowing users to choose their seats during booking.
-- Booking and Reservation: Passengers can book flights and manage their reservations.
-- Admin Dashboard: Airline administrators have access to a dashboard to manage flights, seat availability, and user bookings.
-- Flight Management: Admins can add, update, or remove flights from the system.
-- Reservation Management: Airline staff can manage passenger reservations, including cancellations and modifications.
+- **User Registration and Login** – create accounts and manage your bookings.
+- **Flight Search** – look up available flights by destination and date.
+- **Seat Selection** – choose your preferred seat while booking.
+- **Reservation Management** – modify or cancel reservations.
+- **Admin Dashboard** – manage flights and monitor bookings.
 
 ## Technologies Used
-- Frontend: HTML, CSS, JavaScript
-- Backend: Java Spring Boot
-- Database: MySql
+- Java 17
+- Spring Boot
+- Thymeleaf templates
+- MySQL
+- HTML/CSS/JavaScript
 
-### Clone the Repository
-```bash
-git clone https://github.com/ImeshaDilshani/Airline-Reservation-Management-System.git
-cd Airline-Reservation-Management-System
-```
-### Database Setup
-Create a MySQL database and import the provided schema to set up the necessary tables.
+## Getting Started
+1. Clone this repository:
+   ```bash
+   git clone https://github.com/your-username/myairline-reservation.git
+   cd myairline-reservation
+   ```
+2. Create a MySQL database and update the credentials in `src/main/resources/application.properties`.
+3. Run the application:
+   ```bash
+   ./mvnw spring-boot:run
+   ```
+4. Open your browser to [http://localhost:8080](http://localhost:8080) to access the system.
 
-### Backend Setup
-- Navigate to the backend directory
-- Configure the database connection in the application.properties file.
-
-## Access the Application
-Open your web browser and visit ```http://localhost:8080``` to access the Airline Reservation Management System.
-
-
+---
+This project showcases my exploration of Spring Boot and web development.
