--- src/main/java/com/jkshian/arms/ArmsApplication.java
@@ -1,14 +1,0 @@
-package com.jkshian.arms;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ArmsApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ArmsApplication.class, args);
-    }
-
-
-}
