--- src/main/java/com/jkshian/arms/User/Role.java
@@ -1,6 +1,0 @@
-package com.jkshian.arms.User;
-
-public enum Role {
-    ROLE_USER,
-    ROLE_ADMIN
-}
