--- src/main/java/com/jkshian/arms/dto/AuthenticationRequest.java
@@ -1,15 +1,0 @@
-package com.jkshian.arms.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class AuthenticationRequest {
-    private String email;
-    private String password;
-}
