--- src/main/java/com/jkshian/arms/dto/AuthenticationResponse.java
@@ -1,14 +1,0 @@
-package com.jkshian.arms.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class AuthenticationResponse {
-    private String token;
-}
