--- src/main/java/com/jkshian/arms/dto/RegisterRequest.java
@@ -1,18 +1,0 @@
-package com.jkshian.arms.dto;
-
-import com.jkshian.arms.User.Role;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class RegisterRequest {
-    private String firstname;
-    private String lastname;
-    private String email;
-    private String password;
-}
