--- src/main/java/com/jkshian/arms/entity/AirPlane.java
@@ -1,27 +1,0 @@
-package com.jkshian.arms.entity;
-
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.DynamicUpdate;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-@DynamicUpdate
-@Table(name =  "airplane")
-public class AirPlane {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private  int id;
-    private String start;
-    private String end;
-    private int avlSeat;
-    private double numOfKm;
-}
