--- src/main/java/com/jkshian/arms/entity/Booking.java
@@ -1,27 +1,0 @@
-package com.jkshian.arms.entity;
-
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-@Table(name = "booking")
-public class Booking {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-    private String bStart;
-    private String bEnd;
-    private String userEmail;
-    private int bNumOfseat;
-    private double price;
-
-}
