--- src/main/java/com/jkshian/arms/repo/UserRepository.java
@@ -1,12 +1,0 @@
-package com.jkshian.arms.repo;
-
-
-import com.jkshian.arms.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-@Repository
-public interface UserRepository extends JpaRepository<User,Integer> {
-    Optional<User>findByEmail(String email);
-}
