--- src/main/java/com/jkshian/arms/service/AuthenticationService.java
@@ -1,87 +1,0 @@
-package com.jkshian.arms.service;
-
-import com.jkshian.arms.User.Role;
-import com.jkshian.arms.config.JwtService;
-import com.jkshian.arms.dto.AuthenticationRequest;
-import com.jkshian.arms.dto.AuthenticationResponse;
-import com.jkshian.arms.dto.RegisterRequest;
-import com.jkshian.arms.entity.User;
-import com.jkshian.arms.repo.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class AuthenticationService {
-
-    private final UserRepository repository;
-    private final PasswordEncoder passwordEncoder;
-    private final JwtService jwtService;
-
-
-
-
-    private final AuthenticationManager authenticationManager;
-
-
-    public AuthenticationResponse registerUser(RegisterRequest request) {
-        var user = User.builder()
-                .firstName(request.getFirstname())
-                .lastName(request.getLastname())
-//                .email(passwordEncoder.encode(request.getPassword()))
-                .email(request.getEmail())
-                .password(passwordEncoder.encode(request.getPassword()))
-                .role(Role.ROLE_USER)
-                .build();
-       var exsitUser = repository.findByEmail(user.getEmail());
-       if(exsitUser.isPresent()){
-           return null;
-       }
-        repository.save(user);
-        var jwtToken = jwtService.generateToken(user);
-        return AuthenticationResponse.builder()
-                .token(jwtToken)
-                .build();
-    }
-
-    public AuthenticationResponse registerAdmin(RegisterRequest request) {
-        var user = User.builder()
-                .firstName(request.getFirstname())
-                .lastName(request.getLastname())
-                .email(request.getEmail())
-                .password(passwordEncoder.encode(request.getPassword()))
-                .role(Role.ROLE_ADMIN)
-                .build();
-        var exsitUser = repository.findByEmail(user.getEmail());
-        if(exsitUser.isPresent()){
-            return null;
-        }
-        repository.save(user);
-        var jwtToken = jwtService.generateToken(user);
-        return AuthenticationResponse.builder()
-                .token(jwtToken)
-                .build();
-    }
-
-
-
-    public AuthenticationResponse authenticate(AuthenticationRequest request) {
-        authenticationManager.authenticate(
-                new UsernamePasswordAuthenticationToken(
-                        request.getEmail(),
-                        request.getPassword()
-                )
-        );
-        var user = repository.findByEmail(request.getEmail()).orElseThrow();
-        var jwtToken = jwtService.generateToken(user);
-        return AuthenticationResponse.builder()
-                .token(jwtToken)
-                .build();
-    }
-
-}
-
-
