+++ src/main/java/com/myairline/arms/ArmsApplication.java
@@ -1,0 +1,14 @@
+package com.myairline.arms;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ArmsApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ArmsApplication.class, args);
+    }
+
+
+}
