+++ src/main/java/com/myairline/arms/User/Role.java
@@ -1,0 +1,6 @@
+package com.myairline.arms.User;
+
+public enum Role {
+    ROLE_USER,
+    ROLE_ADMIN
+}
