+++ src/main/java/com/myairline/arms/dto/AuthenticationResponse.java
@@ -1,0 +1,14 @@
+package com.myairline.arms.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class AuthenticationResponse {
+    private String token;
+}
