+++ src/main/java/com/myairline/arms/dto/RegisterRequest.java
@@ -1,0 +1,18 @@
+package com.myairline.arms.dto;
+
+import com.myairline.arms.User.Role;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class RegisterRequest {
+    private String firstname;
+    private String lastname;
+    private String email;
+    private String password;
+}
