+++ src/main/java/com/myairline/arms/entity/AirPlane.java
@@ -1,0 +1,27 @@
+package com.myairline.arms.entity;
+
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.hibernate.annotations.DynamicUpdate;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@DynamicUpdate
+@Table(name =  "airplane")
+public class AirPlane {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private  int id;
+    private String start;
+    private String end;
+    private int avlSeat;
+    private double numOfKm;
+}
