+++ src/main/java/com/myairline/arms/entity/Booking.java
@@ -1,0 +1,27 @@
+package com.myairline.arms.entity;
+
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name = "booking")
+public class Booking {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+    private String bStart;
+    private String bEnd;
+    private String userEmail;
+    private int bNumOfseat;
+    private double price;
+
+}
