+++ src/main/java/com/myairline/arms/repo/UserRepository.java
@@ -1,0 +1,12 @@
+package com.myairline.arms.repo;
+
+
+import com.myairline.arms.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+@Repository
+public interface UserRepository extends JpaRepository<User,Integer> {
+    Optional<User>findByEmail(String email);
+}
