+++ src/main/resources/templates/Admin/flights.html
@@ -2,7 +2,7 @@
 <html lang="en" xmlns="http://www.w3.org/1999/xhtml"  xmlns:th="http://www.w3.org/1999/xhtml" >
 <head>
     <meta charset="UTF-8">
-    <title>JKshian</title>
+    <title>MyAirline</title>
     <!-- Bootstrap core CSS     -->
     <link rel="stylesheet" th:href="@{/css/bootstrap.min.css}"  href="../../static/css/bootstrap.min.css">
     <!-- Animation library for notifications   -->
@@ -21,9 +21,9 @@
     <div class="sidebar" data-color="green" data-image="../static/images/sb2.jpeg" th:data-image="@{/images/sb2.jpeg}" >
         <div class="sidebar-wrapper">
             <div class="logo">
-                <a href="https://github.com/ImeshaDilshani/Airline-Reservation-Management-System.git" class="simple-text">
+                <a href="https://github.com/your-username/myairline-reservation" class="simple-text">
                     <i class="fas fa-plane" aria-hidden="true"></i>
-                    JKshian
+                    MyAirline
                 </a>
             </div>
             <ul class="nav">
@@ -190,7 +190,7 @@
     <footer class="footer">
         <div class="container-fluid">
             <p class="copyright pull-right">
-                &copy; <script>document.write(new Date().getFullYear())</script> <a href="https://github.com/ImeshaDilshani/Airline-Reservation-Management-System.git">JKshian</a>, Connecting People, Places, And Cultures.
+                &copy; <script>document.write(new Date().getFullYear())</script> <a href="https://github.com/your-username/myairline-reservation">MyAirline</a>, Connecting People, Places, And Cultures.
             </p>
         </div>
     </footer>
