+++ src/main/resources/templates/User/contact.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>JKshian</title>
+    <title>MyAirline</title>
     <!------Font awesome cdn link-------->
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.2.1/css/all.min.css">
     <!------CSS file link------->
@@ -15,7 +15,7 @@
 <body>
    <!-- header section starts  -->
    <header class="header">
-      <a href="index.html" class="logo"><i class="fas fa-plane" aria-hidden="true"></i> JKshian</a>
+      <a href="index.html" class="logo"><i class="fas fa-plane" aria-hidden="true"></i> MyAirline</a>
          <nav class="navbar" id="MenuItems">
              <li><a href="index.html">Home</a></li>
              <li><a href="index.html">Booking</a></li>
@@ -106,11 +106,11 @@
     <div class="row">
         <div class="footer-col-1">
          <h3>About us</h3>
-         <p>Launched in 2023, Sri Lanka’s National Carrier is an award winning airline
+         <p>Launched in 2024, Sri Lanka’s National Carrier is an award winning airline
               with a solid reputation for service, comfort, safety, reliability, and punctuality</p>
         </div>
         <div class="footer-col-2">
-         <a href="#" class="logo"><i class="fas fa-plane" aria-hidden="true"></i> JKshian</a>
+         <a href="#" class="logo"><i class="fas fa-plane" aria-hidden="true"></i> MyAirline</a>
         </div>
         <div class="footer-col-3">
             <h3>Useful Links</h3>
@@ -132,7 +132,7 @@
         </div>
     </div>
     <hr>
-    <p class="copyright">Copyright 2023 - JKshian Airlines</p>
+    <p class="copyright">Copyright 2024 - MyAirline Airlines</p>
  </div>
 
 <!-----Footer section ends------->
