+++ src/main/resources/templates/User/index.html
@@ -2,7 +2,7 @@
 <html lang="en"  xmlns="http://www.w3.org/1999/xhtml"  xmlns:th="http://www.w3.org/1999/xhtml">
 <head>
   <meta charset="UTF-8">
-  <title>JKshian</title>
+  <title>MyAirline</title>
   <!------Font awesome cdn link-------->
   <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.2.1/css/all.min.css">
   <!------CSS file link------->
@@ -13,7 +13,7 @@
 <body>
 <!-- header section starts  -->
 <header class="header">
-  <a href="#" class="logo"><i class="fas fa-plane" aria-hidden="true"></i> JKshian</a>
+  <a href="#" class="logo"><i class="fas fa-plane" aria-hidden="true"></i> MyAirline</a>
   <nav class="navbar">
     <a href="index.html">Home</a>
     <a href="#book-form">Booking</a>
@@ -103,7 +103,7 @@
       </div>
     </div>
     <div class="footer-col-2">
-      <a href="#" class="logo"><i class="fas fa-plane" aria-hidden="true"></i> JKshian</a>
+      <a href="#" class="logo"><i class="fas fa-plane" aria-hidden="true"></i> MyAirline</a>
     </div>
     <div class="footer-col-3">
       <h3>Useful Links</h3>
@@ -125,7 +125,7 @@
     </div>
   </div>
   <hr>
-  <p class="copyright">Copyright 2023 - JKshian Airlines</p>
+  <p class="copyright">Copyright 2024 - MyAirline Airlines</p>
 </div>
 ---Footer section ends-----
